(android) Adapt UI for Tor changes
The Tor proxy status UI has been removed. Phoenix now controls
that the Electrum server is an onion address, or shows a message
otherwise. The "enable tor" switch and the confirmation dialog
have been updated.

Some warning about background payments have been amended or
removed since using a Tor VPN app should be more reliable.

Translation have been removed and need to be fixed.
