+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/settings/TorConfigView.kt
@@ -16,9 +16,17 @@
 
 package fr.acinq.phoenix.android.settings
 
-import androidx.compose.foundation.layout.*
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.material.ButtonDefaults
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
@@ -28,120 +36,141 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import fr.acinq.lightning.utils.Connection
+import androidx.compose.ui.window.DialogProperties
+import fr.acinq.phoenix.android.AppViewModel
 import fr.acinq.phoenix.android.R
+import fr.acinq.phoenix.android.application
 import fr.acinq.phoenix.android.business
-import fr.acinq.phoenix.android.components.*
-import fr.acinq.phoenix.android.components.settings.Setting
+import fr.acinq.phoenix.android.components.BorderButton
+import fr.acinq.phoenix.android.components.Button
+import fr.acinq.phoenix.android.components.Card
+import fr.acinq.phoenix.android.components.Checkbox
+import fr.acinq.phoenix.android.components.DefaultScreenHeader
+import fr.acinq.phoenix.android.components.DefaultScreenLayout
+import fr.acinq.phoenix.android.components.Dialog
+import fr.acinq.phoenix.android.components.ProgressView
+import fr.acinq.phoenix.android.components.openLink
 import fr.acinq.phoenix.android.components.settings.SettingSwitch
-import fr.acinq.phoenix.android.navController
 import fr.acinq.phoenix.android.userPrefs
 import fr.acinq.phoenix.android.utils.annotatedStringResource
-import fr.acinq.phoenix.android.utils.negativeColor
-import fr.acinq.phoenix.android.utils.orange
-import fr.acinq.phoenix.android.utils.positiveColor
+import fr.acinq.phoenix.android.utils.mutedBgColor
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 
 @Composable
-fun TorConfigView() {
+fun TorConfigView(
+    appViewModel: AppViewModel,
+    onBackClick: () -> Unit,
+    onBusinessTeardown: () -> Unit,
+) {
     val scope = rememberCoroutineScope()
     val business = business
-    val nc = navController
     val userPrefs = userPrefs
     val torEnabledState = userPrefs.getIsTorEnabled.collectAsState(initial = null)
-    val connState = business.connectionsManager.connections.collectAsState()
 
     var showConfirmTorDialog by remember { mutableStateOf(false) }
 
     DefaultScreenLayout {
         DefaultScreenHeader(
-            onBackClick = { nc.popBackStack() },
+            onBackClick = onBackClick,
             title = stringResource(id = R.string.tor_settings_title),
         )
         val isTorEnabled = torEnabledState.value
         Card {
             if (isTorEnabled == null) {
-                ProgressView(text = stringResource(id = R.string.tor_settings_unknown), space = 12.dp)
+                ProgressView(text = stringResource(id = R.string.utils_loading_prefs), space = 12.dp)
             } else {
                 SettingSwitch(
-                    title = stringResource(id = if (isTorEnabled) R.string.tor_settings_enabled else R.string.tor_settings_disabled),
-                    description = stringResource(id = R.string.tor_settings_subtitle),
+                    title = stringResource(id = R.string.tor_settings_switch_label),
                     icon = R.drawable.ic_tor_shield,
                     enabled = true,
                     isChecked = isTorEnabled,
-                    onCheckChangeAttempt = { enableTor ->
-                        scope.launch {
-                            if (enableTor) {
-                                showConfirmTorDialog = true
-                            } else {
-                                business.appConfigurationManager.updateTorUsage(false)
-                                userPrefs.saveIsTorEnabled(false)
-                            }
-                        }
-                    }
+                    onCheckChangeAttempt = { showConfirmTorDialog = true }
                 )
             }
         }
 
         if (isTorEnabled == true) {
-            Card {
-                val torState = connState.value.tor
-                Setting(
-                    title = when (torState) {
-                        is Connection.CLOSED -> stringResource(R.string.tor_settings_state_closed)
-                        Connection.ESTABLISHING -> stringResource(R.string.tor_settings_state_starting)
-                        Connection.ESTABLISHED -> stringResource(R.string.tor_settings_state_started)
-                        else -> stringResource(R.string.tor_settings_state_unknown)
-                    },
-                    leadingIcon = {
-                        Row(modifier = Modifier.width(width = ButtonDefaults.IconSize), horizontalArrangement = Arrangement.Center) {
-                            Surface(
-                                shape = CircleShape,
-                                color = when (torState) {
-                                    Connection.ESTABLISHED -> positiveColor
-                                    Connection.ESTABLISHING -> orange
-                                    else -> negativeColor
-                                },
-                                modifier = Modifier.size(8.dp)
-                            ) {}
-                        }
-                    }
-                )
+            Card(internalPadding = PaddingValues(16.dp)) {
+                val context = LocalContext.current
+                Text(text = stringResource(id = R.string.tor_settings_instructions_title), style = MaterialTheme.typography.body2)
+                Spacer(Modifier.height(8.dp))
+                Text(text = stringResource(id = R.string.tor_settings_instructions_1))
+                Spacer(Modifier.height(8.dp))
+                Text(text = stringResource(id = R.string.tor_settings_instructions_2))
+                Spacer(Modifier.height(16.dp))
+                BorderButton(text = stringResource(id = R.string.tor_settings_instructions_help_button), icon = R.drawable.ic_external_link, onClick = { openLink(context, "https://phoenix.acinq.co/faq#how-to-use-tor-on-phoenix") })
             }
         }
     }
 
     if (showConfirmTorDialog) {
         var hasReadMessage by remember { mutableStateOf(false) }
+        val isTorEnabled = torEnabledState.value == true
+        val application = application
         Dialog(
             onDismiss = { showConfirmTorDialog = false },
-            title = stringResource(id = R.string.tor_confirm_dialog_title),
-            buttons = {
-                Button(text = stringResource(id = R.string.btn_cancel), onClick = { showConfirmTorDialog = false })
-                Spacer(modifier = Modifier.width(16.dp))
-                Button(
-                    text = stringResource(id = R.string.btn_confirm),
-                    icon = R.drawable.ic_check_circle,
-                    onClick = {
-                        scope.launch {
-                            business.appConfigurationManager.updateTorUsage(true)
-                            userPrefs.saveIsTorEnabled(true)
-                        }
-                        showConfirmTorDialog = false
-                    },
-                    enabled = hasReadMessage
-                )
-            }
+            properties = DialogProperties(dismissOnClickOutside = false, dismissOnBackPress = false, usePlatformDefaultWidth = false),
+            buttons = null
         ) {
-            Column(modifier = Modifier.padding(horizontal = 24.dp)) {
-                Text(text = stringResource(id = R.string.tor_confirm_dialog_details_1))
-                Spacer(modifier = Modifier.height(12.dp))
-                Text(text = annotatedStringResource(id = R.string.tor_confirm_dialog_details_2))
-                Spacer(modifier = Modifier.height(12.dp))
-                Checkbox(text = stringResource(id = R.string.utils_ack), checked = hasReadMessage, onCheckedChange = { hasReadMessage = it })
+            var isTearingDownBusiness by remember { mutableStateOf(false) }
+            if (isTearingDownBusiness) {
+                ProgressView(text = stringResource(id = R.string.tor_dialog_processing), modifier = Modifier.fillMaxWidth(), padding = PaddingValues(32.dp), horizontalArrangement = Arrangement.Center)
+            } else {
+                Column(modifier = Modifier.padding(horizontal = 24.dp)) {
+                    Spacer(modifier = Modifier.height(20.dp))
+                    if (isTorEnabled) {
+                        Text(text = annotatedStringResource(id = R.string.tor_dialog_disable_title), style = MaterialTheme.typography.h4)
+                        Spacer(modifier = Modifier.height(16.dp))
+                        Text(text = annotatedStringResource(id = R.string.tor_dialog_disable_details_1))
+                        Spacer(modifier = Modifier.height(12.dp))
+                        Text(text = annotatedStringResource(id = R.string.tor_dialog_disable_details_2))
+                    } else {
+                        Text(text = stringResource(id = R.string.tor_dialog_enable_title_1), style = MaterialTheme.typography.h4)
+                        Spacer(modifier = Modifier.height(12.dp))
+                        Text(text = stringResource(id = R.string.tor_dialog_enable_details_1))
+                        Spacer(modifier = Modifier.height(16.dp))
+                        Text(text = annotatedStringResource(id = R.string.tor_dialog_enable_title_2), style = MaterialTheme.typography.h4)
+                        Spacer(modifier = Modifier.height(12.dp))
+                        Text(text = annotatedStringResource(id = R.string.tor_dialog_enable_details_2))
+                        Spacer(modifier = Modifier.height(16.dp))
+                        Surface(color = mutedBgColor, shape = RoundedCornerShape(16.dp)) {
+                            Checkbox(text = stringResource(id = R.string.utils_ack), checked = hasReadMessage, onCheckedChange = { hasReadMessage = it }, padding = PaddingValues(16.dp), modifier = Modifier.fillMaxWidth())
+                        }
+                    }
+                }
+                Spacer(modifier = Modifier.width(32.dp))
+                Row(
+                    modifier = Modifier.padding(8.dp).align(Alignment.End)
+                ) {
+                    Button(text = stringResource(id = R.string.btn_cancel), onClick = { showConfirmTorDialog = false }, shape = RoundedCornerShape(16.dp))
+                    Spacer(modifier = Modifier.width(16.dp))
+                    Button(
+                        text = stringResource(id = R.string.btn_confirm),
+                        icon = R.drawable.ic_check_circle,
+                        onClick = {
+                            val service = appViewModel.service ?: return@Button
+                            scope.launch {
+                                isTearingDownBusiness = true
+                                service.shutdown()
+                                application.shutdownBusiness()
+                                business.appConfigurationManager.updateTorUsage(!isTorEnabled)
+                                userPrefs.saveIsTorEnabled(!isTorEnabled)
+                                application.resetBusiness()
+                                delay(500)
+                                showConfirmTorDialog = false
+                                onBusinessTeardown()
+                            }
+                        },
+                        enabled = isTorEnabled || hasReadMessage,
+                        shape = RoundedCornerShape(16.dp)
+                    )
+                }
             }
         }
     }
