+++ phoenix-android/src/main/res/values-b+es+419/important_strings.xml
@@ -92,7 +92,6 @@
     <string name="home_background_restriction_title">Procesamiento en segundo plano restringido</string>.
     <string name="home_background_restriction_body_1">Phoenix no podrá recibir pagos cuando esté en segundo plano, o cuando esté cerrado.</string>
     <string name="home_background_restriction_body_2">Esto ocurre porque:</string>
-    <string name="home_background_restriction_tor">Tor está activado</string>
     <string name="home_background_restriction_powersaver">El dispositivo está en modo de ahorro de energía</string>
     <string name="home_background_restriction_fcm">FCM notificaciones no disponibles</string>
     <string name="home_background_restriction_fcm_details">Si utilizas GrapheneOS o CalyxOS, instala Google Play Services para recibir notificaciones de FCM. Consulta las preguntas frecuentes para obtener más información.</string>
@@ -279,12 +278,6 @@
     <string name="lnurl_withdraw_success">Se envió la solicitud de retiro a <b>%1$s</b>.\n\nPuede pasar algo de tiempo antes de que se envíen los fondos. Entretanto, mantén la aplicación en línea.</string>
 
     <!-- Tor -->
-
-    <string name="tor_settings_subtitle">Phoenix puede usar Tor para proteger tu IP de los demás nodos de la red Lightning y los servidores de Electrum.</string>
-
-    <string name="tor_confirm_dialog_title">¿Habilitar Tor?</string>
-    <string name="tor_confirm_dialog_details_1">Tor es lento y afectará al rendimiento de tus pagos.</string>
-    <string name="tor_confirm_dialog_details_2">Específicamente, <b>no podrá recibir pagos</b> si Phoenix está en segundo plano, o si está cerrado.</string>
 
     <!-- about -->
