+++ phoenix-android/src/main/res/values-pt-rBR/important_strings.xml
@@ -95,7 +95,6 @@
     <string name="home_background_restriction_title">Processamento em segundo plano restrito</string>
     <string name="home_background_restriction_body_1">O Phoenix não poderá receber pagamentos quando estiver em segundo plano ou quando estiver fechado.</string>
     <string name="home_background_restriction_body_2">Isso acontece porque:</string>
-    <string name="home_background_restriction_tor">Tor está ativado</string>
     <string name="home_background_restriction_powersaver">O dispositivo está no modo de economia de energia</string>
     <string name="home_background_restriction_fcm">FCM notificações indisponíveis</string>
     <string name="home_background_restriction_fcm_details">Se você estiver usando GrapheneOS ou CalyxOS, instale o Google Play Services para receber notificações FCM. Consulte as Perguntas frequentes para obter ajuda.</string>
@@ -281,8 +280,6 @@
     <string name="lnurl_withdraw_success">A solicitação de saque foi enviada para <b>%1$s</b>.\n\nPode levar algum tempo até que os fundos sejam enviados. Enquanto isso, mantenha o aplicativo on-line.</string>
 
     <!-- Tor -->
-
-    <string name="tor_settings_subtitle">A Phoenix pode usar o Tor para proteger seu IP dos outros nós da Lightning Network e dos servidores da Electrum.</string>
 
     <!-- about -->
