+++ phoenix-android/src/main/res/values-sw/important_strings.xml
@@ -99,7 +99,6 @@
     <string name="home_background_restriction_title">Usindikaji wa nyuma umepunguzwa</string>
     <string name="home_background_restriction_body_1">Phoenix haitaweza kupokea malipo wakati iko nyuma, au wakati imefungwa.</string>
     <string name="home_background_restriction_body_2">Hii hutokea kwa sababu:</string>
-    <string name="home_background_restriction_tor">Tor imewashwa</string>
     <string name="home_background_restriction_powersaver">Kifaa kipo kwenye hali ya kuokoa nguvu</string>
     <string name="home_background_restriction_fcm">Arifa za FCM hazipatikani</string>
     <string name="home_background_restriction_fcm_details">Kama uko kwenye GrapheneOS au CalyxOS, weka Huduma za Google Play kupata arifa za FCM. Angalia FAQ kwa mwongozo.</string>
@@ -284,12 +283,6 @@
   <string name="lnurl_withdraw_success">Ombi la kutoa fedha limepelekwa kwa <b>%1$s</b>.\n\nInaweza kuchukua muda kabla ya wao kutuma fedha. Tafadhali weka programu kwenye mtandao kwa wakati huu.</string>
 
   <!-- Tor -->
-
-  <string name="tor_settings_subtitle">Phoenix inaweza kutumia Tor kuficha IP yako kutoka kwa nodi zingine kwenye Mtandao wa Lightning na seva za Electrum.</string>
-
-  <string name="tor_confirm_dialog_title">Wezesha Tor?</string>
-  <string name="tor_confirm_dialog_details_1">Tor ni polepole na itaathiri utendaji wa malipo yako.</string>
-  <string name="tor_confirm_dialog_details_2">Hasa, <b>hutoweza kupokea malipo</b> ikiwa Phoenix iko kwenye usuli, au ikiwa imefungwa.</string>
 
   <!-- about -->
