+++ phoenix-android/src/main/res/values-vi/important_strings.xml
@@ -102,7 +102,6 @@
     <string name="home_background_restriction_title">Hạn chế xử lý nền</string>
     <string name="home_background_restriction_body_1">Phoenix sẽ không thể nhận thanh toán khi nó ở chế độ nền hoặc khi nó bị đóng.</string>
     <string name="home_background_restriction_body_2">Điều này xảy ra vì:</string>
-    <string name="home_background_restriction_tor">Tor đã được bật</string>
     <string name="home_background_restriction_powersaver">Thiết bị đang ở chế độ tiết kiệm năng lượng</string>
     <string name="home_background_restriction_fcm">Thông báo FCM không khả dụng</string>
     <string name="home_background_restriction_fcm_details">Nếu bạn đang sử dụng GrapheneOS hoặc CalyxOS, hãy cài đặt Dịch vụ Google Play để nhận thông báo FCM. Hãy xem Câu hỏi thường gặp để được trợ giúp.</string>
@@ -288,12 +287,6 @@
     <string name="lnurl_withdraw_success">Yêu cầu rút tiền đã được gửi tới <b>%1$s</b>.\n\nViệc gửi vốn sẽ mất một chút thời gian. Vui lòng giữ ứng dụng kết nối trực tuyến trong thời gian chờ đợi.</string>
 
     <!-- Tor -->
-
-    <string name="tor_settings_subtitle">Phoenix có thể sử dụng Tor để bảo vệ IP của bạn khỏi các nút mạng khác trên mạng Lightning và máy chủ Electrum.</string>
-
-    <string name="tor_confirm_dialog_title">Bật Tor ?</string>
-    <string name="tor_confirm_dialog_details_1">Tor chạy chậm và sẽ ảnh hưởng đến hiệu suất thanh toán của bạn.</string>
-    <string name="tor_confirm_dialog_details_2">Đặc biệt, bạn <b>sẽ không thể nhận thanh toán</b> nếu Phoenix ở chế độ nền hoặc nếu nó bị đóng.</string>
 
     <!-- about -->
